requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
gunicorn==21.2.0
google-generativeai>=0.5.0
python-dotenv==1.0.0
pydantic==2.5.2
